Move custom code to interfaces_ext.go
